app/database.py
import sqlalchemy
import databases
import os 
from dotenv import load_dotenv, find_dotenv

load_dotenv(find_dotenv())


DATABASE_URL = os.getenv("DATABASE_URL", "sqlite+aiosqlite:///./test.db")

database = databases.Database(DATABASE_URL)

metadata = sqlalchemy.MetaData()

engine = sqlalchemy.create_engine(
    DATABASE_URL.replace("+aiosqlite", ""),
    connect_args={"check_same_thread": False}
)
