requirements.txt
aioboto3==14.1.0
databases==0.9.0
fastapi==0.115.12
jose==1.0.0
pydantic==2.11.3
python-dotenv==1.1.0
python_jose==3.4.0
SQLAlchemy==2.0.40
uvicorn
aiosqlite